requirements.txt
pyautogui
pynput
pillow
opencv-python
opencv-contrib-python
numpy
easyocr
mouse
ahk